Factors out a PDF view for summary and certificate views
