+++ web/views/summary.py
@@ -1,40 +1,17 @@
-import logging
-import os
-
-from django.conf import settings
-from django.http import HttpResponse
-from django.template.loader import get_template
-
-from rest_framework.generics import GenericAPIView
-from xhtml2pdf import pisa
-
 from api.permissions import CanAccessIndividualDeclaration
 from data.models import Declaration, Snapshot
 
-logger = logging.getLogger(__name__)
+from .pdfview import PdfDeclarationView
 
 OTHER_OPTION = "Autre (à préciser)"
 
 
-class SummaryView(GenericAPIView):
+class SummaryView(PdfDeclarationView):
     permission_classes = [CanAccessIndividualDeclaration]
     queryset = Declaration.objects.all()
 
-    def get(self, request, *args, **kwargs):
-        declaration = self.get_object()
-        template = get_template("summary.html")
-        html = template.render(self.get_context(declaration))
-
-        response = HttpResponse(content_type="application/pdf")
-        filename = self.get_pdf_file_name(declaration)
-        response["Content-Disposition"] = f'attachment; filename="{filename}"'
-        pisa_status = pisa.CreatePDF(html, dest=response, link_callback=SummaryView.link_callback)
-
-        if pisa_status.err:
-            logger.error(f"Error while generating PDF for teledeclaration {declaration.id}:\n{pisa_status.err}")
-            return HttpResponse("An error ocurred", status=500)
-
-        return response
+    def get_template_path(self, declaration):
+        return "summary.html"
 
     def get_context(self, declaration):
         product_table_rows = (
@@ -101,19 +78,3 @@ class SummaryView(GenericAPIView):
         # https://github.com/betagouv/complements-alimentaires/issues/1367
         name = declaration.name.encode("ascii", "ignore")
         return f"attestation-{name.decode()}.pdf"
-
-    @staticmethod
-    def link_callback(uri, rel):
-        """
-        Convert HTML URIs to absolute system paths so xhtml2pdf can access those resources.
-        """
-        # Gestion des fichiers STATIC
-        if uri.startswith(settings.STATIC_URL):
-            path = os.path.join(settings.STATIC_ROOT, uri.replace(settings.STATIC_URL, ""))
-        # Gestion des fichiers MEDIA
-        elif uri.startswith(settings.MEDIA_URL):
-            path = os.path.join(settings.MEDIA_ROOT, uri.replace(settings.MEDIA_URL, ""))
-        else:
-            return uri  # On le laisse tel qu'il est car pas static ni media
-
-        return path
